refactor: separating validations into a file of their own
